refactor(plugins): move plugin interface and registry to api/plugins package
